+++ Assets/Scripts/Player/UI player/HealthBar.cs
@@ -8,12 +8,12 @@ public class HealthBar : MonoBehaviour
     public Slider slider;
 
 
-    public void SetMaxHealth(int _health) {
+    public void SetMaxHealth(float _health) {
 
         slider.maxValue = _health;
         slider.value = _health;
     }
-    public void SetHealth(int _health)
+    public void SetHealth(float _health)
     {
         slider.value = _health;
     }
